Rename mouseww to wwmouse
